Rename folder models to model
