+++ src/model/vision.js
@@ -1,0 +1,7 @@
+import {FilesetResolver} from "https://cdn.skypack.dev/@mediapipe/tasks-vision@latest";
+
+const vision = await FilesetResolver.forVisionTasks(
+		"https://cdn.jsdelivr.net/npm/@mediapipe/tasks-vision@latest/wasm",
+);
+
+export default vision;
