--- src/models/handLandmarker.js
@@ -1,13 +1,0 @@
-import {HandLandmarker} from "https://cdn.jsdelivr.net/npm/@mediapipe/tasks-vision@latest";
-import vision from "./vision";
-
-const handModel = await HandLandmarker.createFromOptions(vision, {
-	baseOptions: {
-		modelAssetPath: `https://storage.googleapis.com/mediapipe-models/hand_landmarker/hand_landmarker/float16/1/hand_landmarker.task`,
-		delegate      : "GPU",
-	},
-	runningMode: "VIDEO",
-	numHands   : 2,
-});
-
-export {HandLandmarker as handLandmarks, handModel};
